God commit: Read from Notion CSVs
This changes the playtesting suite to read from Notion CSVs for cards
and companions.
These are the same CSVs that we use to initialize the printed out cards
for playtesting - this allows us to have a single source of truth for
all the cards and companions.
I also changed the source code to be a python wheel for better packaging
:)
This PR also changes the shop and combat to be interactive text-driven
terminals.
This makes the playtesting code much easier to use.