Implement a basic heal in the shop
We want the player to have the option to heal their companions in the
shop.
This should cost a decent amount and should only be usable once.
The player should be encouraged to NOT spend the money on healing and
play more greedily if they can get away with it.