Add Fourth Combat: Bloodseeker
This also tunes the enemies to be more challenging.
The second combat has higher health, so that you do not one shot them
with the team of Aborah and Architect.
The Bloodseeker was surprisingly balanced when Ethan and I playtested
it.
It was scary, but not oppressive.
The game feels relatively balanced, which is good.